improve structure and DRY
