+++ src/components/Product.js
@@ -5,7 +5,7 @@ import { useState } from "react";
 import InvestButton from "./InvestButton";
 
 export default function Product() {
-  const { setMoney, tiers } = useStore();
+  const { setMoney, tiers, setTier, money } = useStore();
   const [isFilling, setIsFilling] = useState(false);
 
   // in preperation for multiple tiers
@@ -23,6 +23,22 @@ export default function Product() {
     setMoney(currentTier.income);
   }
 
+  function handleInvest() {
+    const currentTier = getTierById("tier1");
+    if (currentTier.investPrice > money) {
+      return;
+    }
+    setMoney(-currentTier.investPrice);
+
+    setTier({
+      id: "tier1",
+      income: currentTier.income + 5,
+      investPrice: currentTier.investPrice + 50,
+    });
+  }
+
+  const { investPrice } = getTierById("tier1");
+
   return (
     <>
       <MoneyButton
@@ -32,7 +48,11 @@ export default function Product() {
         onTimerEnd={handleTimerEnd}
       />
       <ProgressBar isFilling={isFilling} tier={tiers[0]} />
-      <InvestButton />
+      <InvestButton
+        onInvest={handleInvest}
+        money={money}
+        investPrice={investPrice}
+      />
     </>
   );
 }
